+++ src/transactions/manage_asset/CreateAssetOpFrame.cpp
@@ -204,11 +204,6 @@ vector<longstring> CreateAssetOpFrame::makeTasksKeyVector(StorageHelper& storage
 
 bool CreateAssetOpFrame::ensureUpdateRequestValid(ReviewableRequestFrame::pointer request)
 {
-    if (request->getRejectReason().empty()) {
-        innerResult().code(ManageAssetResultCode::PENDING_REQUEST_UPDATE_NOT_ALLOWED);
-        return false;
-    }
-
     if (mManageAsset.request.createAssetCreationRequest().allTasks)
     {
         innerResult().code(ManageAssetResultCode::NOT_ALLOWED_TO_SET_TASKS_ON_UPDATE);